updated with save and remove

+++ routes/api/saveconfig.js
@@ -3,14 +3,24 @@ const router = express.Router();
 const { check, validationResult } = require('express-validator');
 const CountryLang = require('../../models/CountryLang');
 const auth = require('../../middleware/auth');
+
 // @rout    GET api/user
 // @desc    Test countries route
 // @access  Private
 //router.get('/', (req, res) => res.send('Lang route'));
 
-router.get('/', auth, async (req, res) => {
+router.post('/', auth, async (req, res) => {
 	try {
-		await CountryLang.findAll().then((countries)=>{
+		if(!req.body.countrycode){
+			return res.status(400).json({ errors: 'Invalid CountryCode' });
+		}	
+		await CountryLang.findAll(
+			{
+				where :{
+					countrycode : req.body.countrycode
+				}
+			}
+		).then((countries)=>{
             res.status(200).json(countries);
         })
 	} catch (error) {
@@ -48,14 +58,21 @@ router.post('/save', auth , async(req,res)=>{
 	}
 });
 router.delete('/delete', auth, async(req,res)=>{
-	//var countrycode = req.body.countrycode;
+	var countrycode = req.body.countrycode;
 	var locale = req.body.locale;
 	var constraint = req.body.configlist;
 	try{
+		if(!req.body.locale){
+			return res.status(400).json({ errors: 'Please select locale' });
+		}
+		if(!req.body.countrycode){
+			return res.status(400).json({ errors: 'Invalid CountryCode' });
+		}	
 		await CountryLang.destroy({
 			where :{
 				locale : locale,
-				setlocale: constraint
+				setlocale: constraint,
+				countrycode : countrycode
 			}
 		})
 		.then((response)=>res.status(200).json(response));
